requirements.txt
flask>=3.0.0
werkzeug>=3.0.0
pywebview>=4.0.0
pyinstaller>=6.0.0
